+++ lib/src/dataservice/dataservice.dart
@@ -1,0 +1,8 @@
+library dataservice;
+
+import 'package:timeseries/timeseries.dart';
+import 'package:timezone/timezone.dart';
+
+abstract class DataService {
+  Future<TimeSeries<num>> get(String id, {required Location location});
+}