+++ lib/src/juice_return.dart
@@ -1,0 +1,5 @@
+class JuiceReturn {
+  final Object? value;
+
+  JuiceReturn(this.value);
+}