+++ lib/src/timeseries/time_filter.dart
@@ -1,0 +1,10 @@
+import 'package:elec/elec.dart';
+
+class TimeFilter {
+  TimeFilter();
+
+  Bucket? bucket;
+  String? hours;
+  String? monthRange;
+  String? holiday;
+}