+++ CHANGELOG.md
@@ -1,3 +1,8 @@
-## 1.0.0
+## 0.0.4
+- Add a `runMode` to the `Juice` class.  In debug mode, you can capture 
+  the output of an error during testing. 
+- Removed all references of Lox.
+- Fix reading of multi lines in scripts.  
 
+## 0.0.3
 - Initial version.